LLM-Pedigree-Builder/requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-ollama>=0.1.0
networkx>=3.3
matplotlib>=3.8.0
qrcode>=7.4.2
reportlab>=4.2.0
opencv-python>=4.9.0.80
pyzbar>=0.1.9
pytesseract>=0.3.10
Pillow>=10.3.0
numpy>=1.26.4
pytest>=8.2.0
pytest-cov>=5.0.0
speechrecognition>=3.10.4
gTTS>=2.5.1
